+++ myems-api/core/menu.py
@@ -70,18 +70,6 @@ class MenuCollection:
         pass
 
     @staticmethod
-    def on_options(req, resp):
-        """
-        Handle OPTIONS request for CORS preflight
-
-        Args:
-            req: Falcon request object
-            resp: Falcon response object
-        """
-        _ = req
-        resp.status = falcon.HTTP_200
-
-    @staticmethod
     def on_get(req, resp):
         """
         Handle GET requests to retrieve all menus
@@ -188,20 +176,6 @@ class MenuItem:
 
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_):
-        """
-        Handle OPTIONS request for CORS preflight
-
-        Args:
-            req: Falcon request object
-            resp: Falcon response object
-            id_: Menu ID parameter
-        """
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_):
@@ -384,20 +358,6 @@ class MenuChildrenCollection:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_):
-        """
-        Handle OPTIONS request for CORS preflight
-
-        Args:
-            req: Falcon request object
-            resp: Falcon response object
-            id_: Menu ID parameter
-        """
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-
-    @staticmethod
     def on_get(req, resp, id_):
         """
         Handle GET requests to retrieve menu hierarchy
@@ -545,18 +505,6 @@ class MenuWebCollection:
 
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp):
-        """
-        Handle OPTIONS request for CORS preflight
-
-        Args:
-            req: Falcon request object
-            resp: Falcon response object
-        """
-        _ = req
-        resp.status = falcon.HTTP_200
 
     @staticmethod
     def on_get(req, resp):