+++ myems-api/core/photovoltaicpowerstation.py
@@ -21,18 +21,6 @@ class PhotovoltaicPowerStationCollection:
         pass
 
     @staticmethod
-    def on_options(req, resp):
-        """
-        Handle OPTIONS request for CORS preflight
-
-        Args:
-            req: Falcon request object
-            resp: Falcon response object
-        """
-        _ = req
-        resp.status = falcon.HTTP_200
-
-    @staticmethod
     def on_get(req, resp):
         """
         Handle GET requests to retrieve all photovoltaic power stations
@@ -340,12 +328,6 @@ class PhotovoltaicPowerStationCollection:
 class PhotovoltaicPowerStationItem:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_):
@@ -692,12 +674,6 @@ class PhotovoltaicPowerStationDataSourcePointCollection:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-
-    @staticmethod
     def on_get(req, resp, id_):
         if 'API-KEY' not in req.headers or \
                 not isinstance(req.headers['API-KEY'], str) or \
@@ -745,12 +721,6 @@ class PhotovoltaicPowerStationDataSourcePointCollection:
 class PhotovoltaicPowerStationExport:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_):
@@ -846,11 +816,6 @@ class PhotovoltaicPowerStationExport:
 class PhotovoltaicPowerStationImport:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp):
-        _ = req
-        resp.status = falcon.HTTP_200
 
     @staticmethod
     @user_logger
@@ -1052,12 +1017,6 @@ class PhotovoltaicPowerStationClone:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-
-    @staticmethod
     @user_logger
     def on_post(req, resp, id_):
         admin_control(req)
@@ -1145,12 +1104,6 @@ class PhotovoltaicPowerStationClone:
 class PhotovoltaicPowerStationGridCollection:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_):
@@ -1528,13 +1481,6 @@ class PhotovoltaicPowerStationGridCollection:
 class PhotovoltaicPowerStationGridItem:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, gid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = gid
 
     @staticmethod
     def on_get(req, resp, id_, gid):
@@ -1980,12 +1926,6 @@ class PhotovoltaicPowerStationGridItem:
 class PhotovoltaicPowerStationInvertorCollection:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_):
@@ -3093,12 +3033,6 @@ class PhotovoltaicPowerStationInvertorCollection:
 class PhotovoltaicPowerStationInvertorItem:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, iid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_, iid):
@@ -4243,12 +4177,6 @@ class PhotovoltaicPowerStationLoadCollection:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-
-    @staticmethod
     def on_get(req, resp, id_):
         access_control(req)
         if not id_.isdigit() or int(id_) <= 0:
@@ -4602,12 +4530,6 @@ class PhotovoltaicPowerStationLoadCollection:
 class PhotovoltaicPowerStationLoadItem:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, lid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_, lid):
@@ -5030,12 +4952,6 @@ class PhotovoltaicPowerStationUserCollection:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-
-    @staticmethod
     def on_get(req, resp, id_):
         access_control(req)
         if not id_.isdigit() or int(id_) <= 0:
@@ -5168,12 +5084,6 @@ class PhotovoltaicPowerStationUserItem:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_, uid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-
-    @staticmethod
     @user_logger
     def on_delete(req, resp, id_, uid):
         # todo Verify if the user is bound when deleting it
@@ -5240,12 +5150,6 @@ class PhotovoltaicPowerStationUserItem:
 class PhotovoltaicPowerStationDataSourceCollection:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
 
     @staticmethod
     def on_get(req, resp, id_):
@@ -5380,13 +5284,6 @@ class PhotovoltaicPowerStationDataSourceItem:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_, dsid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = dsid
-
-    @staticmethod
     @user_logger
     def on_delete(req, resp, id_, dsid):
         admin_control(req)
@@ -5443,13 +5340,6 @@ class PhotovoltaicPowerStationDataSourceItem:
 class PhotovoltaicPowerStationInvertorPointCollection:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, iid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = iid
 
     @staticmethod
     def on_get(req, resp, id_, iid):
@@ -5584,14 +5474,6 @@ class PhotovoltaicPowerStationInvertorPointItem:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_, iid, pid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = iid
-        _ = pid
-
-    @staticmethod
     @user_logger
     def on_delete(req, resp, id_, iid, pid):
         admin_control(req)
@@ -5650,13 +5532,6 @@ class PhotovoltaicPowerStationInvertorPointItem:
 class PhotovoltaicPowerStationGridPointCollection:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, gid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = gid
 
     @staticmethod
     def on_get(req, resp, id_, gid):
@@ -5791,14 +5666,6 @@ class PhotovoltaicPowerStationGridPointItem:
         pass
 
     @staticmethod
-    def on_options(req, resp, id_, gid, pid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = gid
-        _ = pid
-
-    @staticmethod
     @user_logger
     def on_delete(req, resp, id_, gid, pid):
         admin_control(req)
@@ -5857,13 +5724,6 @@ class PhotovoltaicPowerStationGridPointItem:
 class PhotovoltaicPowerStationLoadPointCollection:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, lid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = lid
 
     @staticmethod
     def on_get(req, resp, id_, lid):
@@ -5996,14 +5856,6 @@ class PhotovoltaicPowerStationLoadPointCollection:
 class PhotovoltaicPowerStationLoadPointItem:
     def __init__(self):
         pass
-
-    @staticmethod
-    def on_options(req, resp, id_, lid, pid):
-        _ = req
-        resp.status = falcon.HTTP_200
-        _ = id_
-        _ = lid
-        _ = pid
 
     @staticmethod
     @user_logger