+++ myems-api/reports/combinedequipmentefficiency.py
@@ -53,11 +53,6 @@ class Reporting:
         """"Initializes Reporting"""
         pass
 
-    @staticmethod
-    def on_options(req, resp):
-        _ = req
-        resp.status = falcon.HTTP_200
-
     ####################################################################################################################
     # PROCEDURES
     # Step 1: valid parameters