+++ myems-api/reports/equipmentprediction.py
@@ -33,11 +33,6 @@ class Reporting:
         """Initializes Class"""
         pass
 
-    @staticmethod
-    def on_options(req, resp):
-        _ = req
-        resp.status = falcon.HTTP_200
-
     ####################################################################################################################
     # PROCEDURES
     # Step 1: valid parameters